requirements.txt
alembic==1.12.1
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
numpy==1.25.1
pandas==2.0.3
pdf2image==1.16.3
Pillow==10.0.0
psycopg2-binary==2.9.9
pytesseract==0.3.10
SQLAlchemy==2.0.23
tabula-py==2.8.2
python-dotenv==1.0.0
